requirements.txt
requests~=2.32.3
beautifulsoup4~=4.13.4
selenium~=4.32.0
google-generativeai
openai~=1.77.0
dotenv~=0.9.9
python-dotenv~=1.1.0
pyperclip~=1.9.0
wxPython~=4.2.3
pillow~=11.2.1
pyautogui~=0.9.54
moviepy~=2.2.1
protobuf~=5.29.5
pywin32
pywin32-ctypes
clipboard~=0.0.4
matplotlib~=3.10.3

webdriver-manager~=4.0.2